// src/utils/types/mud.ts
export interface FluidData {
  type: string;
  baseFluid: string;
  activePitsVolume: string;
  flowlineTemp: string;
  rheologySource: string;
  pv: string;
  yp: string;
  gel10s: string;
  gel10m: string;
  oilWater: string;
  salinity: string;
  surfaceTemp: string;
  bottomholeTemp: string;
  tempGradient: string;
  gasSolubility: string;
  compressibilityFactor: string;
  gasOilRatio: string;
  viscometerCalDate: string;
  densityCalDate: string;
  tempSensorOffset: string;
}